Move TlsConfiguration under client::options
